fix(cves): cast SUM results to BIGINT to prevent NUMERIC type mismatch
SUM(BIGINT) in PostgreSQL returns NUMERIC, not BIGINT. The inner
COUNT(DISTINCT ...)::BIGINT cast on CTE columns does not carry through
SUM() so the outer COALESCE(SUM(...), 0) resolves to NUMERIC and sqlx
fails to decode it into i64.

Add ::BIGINT casts to all four COALESCE(SUM(...), 0) expressions in
fetch_cve_summary (queries/dashboard.rs) to match what
CVE_DASHBOARD_SUMMARY_SQL already does correctly.

Also parenthesise COUNT(*) FILTER (...) before ::BIGINT in
CVE_DASHBOARD_SUMMARY_SQL so operator precedence is unambiguous.

Fixes runtime 500: 'Rust type i64 is not compatible with SQL type NUMERIC'

+++ packages/default/src/queries/dashboard.rs
@@ -92,10 +92,10 @@ pub async fn fetch_cve_summary(pool: &PgPool) -> Result<CveSummary> {
             GROUP BY v.hostname \
          ) \
          SELECT \
-            COALESCE(SUM(critical_cves), 0), \
-            COALESCE(SUM(high_cves), 0), \
-            COALESCE(SUM(medium_cves), 0), \
-            COALESCE(SUM(low_cves), 0) \
+            COALESCE(SUM(critical_cves), 0)::BIGINT, \
+            COALESCE(SUM(high_cves), 0)::BIGINT, \
+            COALESCE(SUM(medium_cves), 0)::BIGINT, \
+            COALESCE(SUM(low_cves), 0)::BIGINT \
          FROM per_system_counts",
     )
     .fetch_one(pool)
